blogsite/requirements.txt
aniso8601==9.0.1
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgi-tools==0.64.1
asgiref==3.5.2
astroid==2.11.6
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
better-profanity==0.7.0
black==19.3b0
bleach==5.0.0
blinker==1.4
blis==0.7.8
Brlapi==0.7.0
cached-property==1.5.2
catalogue==1.0.0
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.0
chrome-gnome-shell==0.0.0
click==8.1.3
colorama==0.4.5
command-not-found==0.3
coverage==6.3.2
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.6
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.6
django-crispy-forms==1.14.0
djangorestframework==3.13.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
executing==0.8.3
Faker==13.15.1
fasteners==0.14.1
fastjsonschema==2.15.3
filelock==3.6.0
flake8==4.0.1
Flask==2.1.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
fonttools==4.32.0
future==0.18.2
greenlet==1.1.2
h11==0.13.0
http-router==2.6.5
httplib2==0.14.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.0
iniconfig==1.1.1
install==1.3.5
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keyring==18.0.1
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
lockfile==0.12.2
logilab-common==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
modconfig==1.2.1
monotonic==1.5
more-itertools==8.13.0
motor==3.0.0
muffin==0.87.1
muffin-mongo==0.4.0
multidict==6.0.2
murmurhash==1.0.7
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
netifaces==0.10.4
nose==1.3.7
notebook==6.4.10
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
ordered-set==3.1.1
ordered-set-stubs==0.1.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipreqs==0.4.11
plac==1.1.3
platformdirs==2.5.2
pluggy==1.0.0
poetry-version==0.1.5
preshed==3.0.6
profanity-filter==1.3.3
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.6.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycups==1.9.73
pydantic==1.9.1
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.14.3
pymacaroons==0.13.0
pymongo==4.2.0
PyNaCl==1.3.0
pyparsing==3.0.8
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.1.1
pytest-cov==3.0.0
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
redis==3.5.3
reportlab==3.5.34
requests==2.28.1
requests-unixsocket==0.2.0
ruamel.yaml==0.15.100
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==2.3.7
SQLAlchemy==1.4.36
sqlparse==0.4.2
srsly==1.0.5
stack-data==0.2.0
systemd-python==234
terminado==0.13.3
thinc==7.4.5
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.0
tomli==2.0.1
tomlkit==0.5.11
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing-extensions==4.2.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.4.0
unattended-upgrades==0.1
urllib3==1.26.11
usb-creator==0.3.7
uvicorn==0.18.2
virtualenv==20.14.1
wadllib==1.3.3
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.14.1
xgboost==1.6.0
xkit==0.0.0
yarg==0.1.9
yarl==1.7.2
zipp==3.8.0
aniso8601==9.0.1
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgi-tools==0.64.1
asgiref==3.5.2
astroid==2.11.6
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
better-profanity==0.7.0
black==19.3b0
bleach==5.0.0
blinker==1.4
blis==0.7.8
Brlapi==0.7.0
cached-property==1.5.2
catalogue==1.0.0
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.0
chrome-gnome-shell==0.0.0
click==8.1.3
colorama==0.4.5
command-not-found==0.3
coverage==6.3.2
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.6
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.6
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.13.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
executing==0.8.3
Faker==13.15.1
fasteners==0.14.1
fastjsonschema==2.15.3
filelock==3.6.0
flake8==4.0.1
Flask==2.1.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
fonttools==4.32.0
future==0.18.2
greenlet==1.1.2
h11==0.13.0
http-router==2.6.5
httplib2==0.14.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.0
iniconfig==1.1.1
install==1.3.5
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keyring==18.0.1
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
lockfile==0.12.2
logilab-common==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
modconfig==1.2.1
monotonic==1.5
more-itertools==8.13.0
motor==3.0.0
muffin==0.87.1
muffin-mongo==0.4.0
multidict==6.0.2
murmurhash==1.0.7
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
netifaces==0.10.4
nose==1.3.7
notebook==6.4.10
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
ordered-set==3.1.1
ordered-set-stubs==0.1.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipreqs==0.4.11
plac==1.1.3
platformdirs==2.5.2
pluggy==1.0.0
poetry-version==0.1.5
preshed==3.0.6
profanity-filter==1.3.3
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.6.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycups==1.9.73
pydantic==1.9.1
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.14.3
pymacaroons==0.13.0
pymongo==4.2.0
PyNaCl==1.3.0
pyparsing==3.0.8
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.1.1
pytest-cov==3.0.0
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
redis==3.5.3
reportlab==3.5.34
requests==2.28.1
requests-unixsocket==0.2.0
ruamel.yaml==0.15.100
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==2.3.7
SQLAlchemy==1.4.36
sqlparse==0.4.2
srsly==1.0.5
stack-data==0.2.0
systemd-python==234
terminado==0.13.3
thinc==7.4.5
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.0
tomli==2.0.1
tomlkit==0.5.11
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing-extensions==4.2.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.4.0
unattended-upgrades==0.1
urllib3==1.26.11
usb-creator==0.3.7
uvicorn==0.18.2
virtualenv==20.14.1
wadllib==1.3.3
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.14.1
xgboost==1.6.0
xkit==0.0.0
yarg==0.1.9
yarl==1.7.2
zipp==3.8.0
